+++ packages/frontend/src/components/Explorer/CustomMetricModal/index.tsx
@@ -1,17 +1,9 @@
 import {
-    ConditionalOperator,
-    createFilterRuleFromField,
     Dimension,
-    Field,
-    FieldTarget,
-    FilterRule,
     friendlyName,
-    isField,
-    isFilterableField,
     MetricFilterRule,
     MetricType,
     snakeCaseName,
-    TableCalculation,
 } from '@lightdash/common';
 import {
     Accordion,
@@ -27,18 +19,10 @@ import { useParams } from 'react-router-dom';
 import { useExplore } from '../../../hooks/useExplore';
 import { useProject } from '../../../hooks/useProject';
 import { useExplorerContext } from '../../../providers/ExplorerProvider';
-import FilterRuleForm from '../../common/Filters/FilterRuleForm';
-import {
-    FiltersProvider,
-    useFiltersContext,
-} from '../../common/Filters/FiltersProvider';
+import { FiltersProvider } from '../../common/Filters/FiltersProvider';
 import { useFieldsWithSuggestions } from '../FiltersCard/useFieldsWithSuggestions';
-
-interface MetricFilterRuleWithFieldIds
-    extends FilterRule<
-        ConditionalOperator,
-        FieldTarget & { fieldRef: string }
-    > {}
+import { FilterForm, MetricFilterRuleWithFieldId } from './FilterForm';
+import { addFieldIdToMetricFilterRule } from './utils';
 
 type Props = {
     isEditMode: boolean;
@@ -48,117 +32,13 @@ type Props = {
     item: Dimension;
 };
 
-const FilterForm: FC<{
-    item: Dimension;
-    customMetricFiltersWithIds: MetricFilterRuleWithFieldIds[];
-    setCustomMetricFiltersWithIds: Dispatch<
-        SetStateAction<MetricFilterRuleWithFieldIds[]>
-    >;
-}> = ({ item, customMetricFiltersWithIds, setCustomMetricFiltersWithIds }) => {
-    const isEditMode = useExplorerContext(
-        (context) => context.state.isEditMode,
-    );
-
-    const { fieldsMap } = useFiltersContext();
-
-    const fields = Object.values(fieldsMap);
-
-    const addFieldRule = useCallback(
-        (field: Field | TableCalculation | Dimension) => {
-            if (isField(field) && isFilterableField(field)) {
-                const newFilterRule = createFilterRuleFromField(field);
-
-                setCustomMetricFiltersWithIds([
-                    ...customMetricFiltersWithIds,
-                    {
-                        ...newFilterRule,
-                        target: {
-                            ...newFilterRule.target,
-                            fieldRef: `${field.table}.${field.name}`,
-                        },
-                    },
-                ]);
-            }
-        },
-        [customMetricFiltersWithIds, setCustomMetricFiltersWithIds],
-    );
-
-    const onChangeItem = useCallback(
-        (itemIndex: number, filterRule: FilterRule) => {
-            setCustomMetricFiltersWithIds(
-                customMetricFiltersWithIds.map((customMetricFilter, index) =>
-                    itemIndex === index
-                        ? {
-                              ...filterRule,
-                              target: {
-                                  ...filterRule.target,
-                                  fieldRef: `${
-                                      fieldsMap[filterRule.target.fieldId].table
-                                  }.${
-                                      fieldsMap[filterRule.target.fieldId].name
-                                  }`,
-                              },
-                          }
-                        : customMetricFilter,
-                ),
-            );
-        },
-        [customMetricFiltersWithIds, fieldsMap, setCustomMetricFiltersWithIds],
-    );
-
-    const onDeleteItem = useCallback(
-        (index: number) => {
-            setCustomMetricFiltersWithIds(
-                customMetricFiltersWithIds.filter((value, i) => i !== index),
-            );
-        },
-        [customMetricFiltersWithIds, setCustomMetricFiltersWithIds],
-    );
-
-    return (
-        <Stack spacing="sm">
-            {customMetricFiltersWithIds.map((filterRule, index) => (
-                <FilterRuleForm
-                    key={filterRule.id}
-                    filterRule={filterRule}
-                    fields={fields}
-                    isEditMode={isEditMode}
-                    onChange={(value) => onChangeItem(index, value)}
-                    onDelete={() => onDeleteItem(index)}
-                />
-            ))}
-            <Button
-                display="block"
-                mr="auto"
-                size="xs"
-                variant="outline"
-                onClick={() => {
-                    addFieldRule(item);
-                }}
-                disabled={fields.length <= 0}
-            >
-                Add filter
-            </Button>
-        </Stack>
-    );
-};
-
-export const CreateCustomMetricModal: FC<Props> = ({
+export const CustomMetricModal: FC<Props> = ({
     isEditMode,
     item,
     isCreatingCustomMetric,
     setIsCreatingCustomMetric,
     customMetricType,
 }) => {
-    const [customMetricName, setCustomMetricName] = useState(
-        customMetricType
-            ? `${friendlyName(customMetricType)} of ${item.label}`
-            : '',
-    );
-
-    const [customMetricFiltersWithIds, setCustomMetricFiltersWithIds] =
-        useState<MetricFilterRuleWithFieldIds[]>([]);
-
     const { projectUuid } = useParams<{ projectUuid: string }>();
     const project = useProject(projectUuid);
 
@@ -169,21 +49,49 @@ export const CreateCustomMetricModal: FC<Props> = ({
         (context) => context.state.unsavedChartVersion.tableName,
     );
 
-    const additionalMetrics = useExplorerContext(
-        (context) =>
-            context.state.unsavedChartVersion.metricQuery.additionalMetrics,
-    );
     const queryResults = useExplorerContext(
         (context) => context.queryResults.data,
     );
 
     const { data } = useExplore(tableName);
 
+    const additionalMetrics = useExplorerContext(
+        (context) =>
+            context.state.unsavedChartVersion.metricQuery.additionalMetrics,
+    );
+
     const fieldsWithSuggestions = useFieldsWithSuggestions({
         data,
         queryResults,
         additionalMetrics,
     });
+
+    const [customMetricName, setCustomMetricName] = useState(
+        isEditMode
+            ? item.label
+            : customMetricType
+            ? `${friendlyName(customMetricType)} of ${item.label}`
+            : '',
+    );
+
+    const getCurrentCustomMetricFiltersWithIds = useCallback(() => {
+        const currentCustomMetric = additionalMetrics?.find(
+            (metric) => metric.label === customMetricName,
+        );
+
+        if (currentCustomMetric && currentCustomMetric.filters) {
+            return currentCustomMetric.filters.map((filterRule) =>
+                addFieldIdToMetricFilterRule(filterRule),
+            );
+        }
+
+        return [];
+    }, [additionalMetrics, customMetricName]);
+
+    const [customMetricFiltersWithIds, setCustomMetricFiltersWithIds] =
+        useState<MetricFilterRuleWithFieldId[]>(
+            isEditMode ? getCurrentCustomMetricFiltersWithIds() : [],
+        );
 
     const createCustomMetric = useCallback(
         (dimension: Dimension, type: MetricType) => {
@@ -234,6 +142,7 @@ export const CreateCustomMetricModal: FC<Props> = ({
                 ...(customMetricFilters.length > 0 && {
                     filters: customMetricFilters,
                 }),
+                baseFieldId: dimension.name,
                 ...format,
                 ...round,
                 ...compact,
